refactor(userinfo): 优化用户账户状态校验逻辑
- 拆分生命周期状态和封禁状态多重校验为独立方法
- 永久封禁直接拒绝登录，临时封禁超过过期时间自动放行
- 增加对未知封禁类型的容错处理和日志记录
- 调整账户锁定时的登录失败记录逻辑和异常抛出
- 统一提升代码可读性和维护性

+++ ydsz-cronjob/ydsz-cronjob-infra/src/main/java/com/njydsz/cronjob/infra/repository/impl/JobTaskRepositoryImpl.java
@@ -3,11 +3,11 @@ package com.njydsz.cronjob.infra.repository.impl;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
-import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
-import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.njydsz.cronjob.domain.repository.JobRepository;
 import com.njydsz.cronjob.domain.repository.JobTaskRepository;
 import com.njydsz.cronjob.domain.vo.JobTaskVO;
@@ -76,7 +76,7 @@ public class JobTaskRepositoryImpl implements JobTaskRepository {
   public int countByLogId(String logId) {
     LambdaQueryWrapper<JobTask> wrapper = new LambdaQueryWrapper<>();
     wrapper.eq(JobTask::getLogId, logId).eq(JobTask::getDeleted, 0);
-    return jobTaskMapper.selectCount(wrapper);
+    return jobTaskMapper.selectCount(wrapper).intValue();
   }
 
   @Override
